Organize article scripts and add module summaries
